% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{boot}
\alias{boot}
\title{Non-parametric Bootstrapping}
\usage{
boot(data, func = NULL, B)
}
\arguments{
\item{data}{he data as a vector, matrix or data frame}

\item{func}{function to be bootstrapped}

\item{B}{the number of replicates}
}
\value{
a list of standard error and bias
}
\description{
Estimate the standard error and the bias of an estimator using R
}
\examples{
\dontrun{
data <- 20 * rbeta(1000,2,3)
boot(data = data, func = mean, B = 2000)
}
}
