% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{jack}
\alias{jack}
\title{Non-parametric Bootstrapping}
\usage{
jack(data, func = NULL)
}
\arguments{
\item{data}{he data as a vector}

\item{func}{function to be bootstrapped}
}
\value{
the standard error to be estimated
}
\description{
Compute the jackknife estimate of standard error using R
}
\examples{
\dontrun{
data <- 20 * rbeta(1000,2,3)
jack(data = data, func = mean)
}
}
