% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{jackafterboot}
\alias{jackafterboot}
\title{Non-parametric Bootstrapping}
\usage{
jackafterboot(data, func = NULL, B)
}
\arguments{
\item{data}{the data as a vector}

\item{func}{function to be bootstrapped}

\item{B}{the number of replicates}
}
\value{
the standard error of both bootstrap method and jackknife-after-bootstrap method
}
\description{
Computes an estimate for each leave-one-out sample using R
}
\examples{
\dontrun{
data <- 20 * rbeta(1000,2,3)
jackafterboot(data = data, func=mean, B = 2000)
}
}
