\name{StatComp20036-package}
\alias{StatComp20036-package}
\alias{StatComp20036}
\docType{package}
\title{
  This package includes several functions for boostrapping and kNN using Rcpp.
}
\description{
  Several self-implemented functions of boot, jack, jackafterboot and kNN are provided in this package {\code{boot} and \code{jack} and \code{jackafterboot} and \code{kNN}}
}
\author{
Can Wang

Maintainer: Can Wang <wcancan@mail.ustc.edu.cn>
}

\keyword{ package }
\seealso{
  Optional links to other man pages
}